Switch group offset to uint32 across APIs

+++ lib/api/src/grpc/qdrant.rs
@@ -5023,8 +5023,8 @@ pub struct SearchPointGroups {
     #[prost(message, optional, tag = "16")]
     pub sparse_indices: ::core::option::Option<SparseIndices>,
     /// Offset of the first group to return
-    #[prost(uint64, optional, tag = "17")]
-    pub offset: ::core::option::Option<u64>,
+    #[prost(uint32, optional, tag = "17")]
+    pub offset: ::core::option::Option<u32>,
 }
 #[derive(serde::Serialize)]
 #[allow(clippy::derive_partial_eq_without_eq)]
@@ -5287,8 +5287,8 @@ pub struct RecommendPointGroups {
     #[prost(message, optional, tag = "21")]
     pub shard_key_selector: ::core::option::Option<ShardKeySelector>,
     /// Offset of the first group to return
-    #[prost(uint64, optional, tag = "22")]
-    pub offset: ::core::option::Option<u64>,
+    #[prost(uint32, optional, tag = "22")]
+    pub offset: ::core::option::Option<u32>,
 }
 #[derive(serde::Serialize)]
 #[allow(clippy::derive_partial_eq_without_eq)]
